# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Standard email and password login",
        "description": "Authenticates system user or admin and returns a signed JWT bearer token.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated successfully; returns session token."
          },
          "401": {
            "description": "Invalid email or password."
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "summary": "Current User Identity",
        "description": "Protected test route to verify your JWT token in Swagger.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current authenticated user profile and roles."
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/invitations/validate": {
      "get": {
        "operationId": "InvitationsController_validateToken",
        "summary": "Validate invite token status before rendering the claim view",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token details and validation status"
          },
          "400": {
            "description": "Token has already been consumed or expired"
          },
          "404": {
            "description": "Token not found"
          }
        },
        "tags": [
          "Invitations & Onboarding"
        ]
      }
    },
    "/auth/claim-owner-invite": {
      "post": {
        "operationId": "InvitationsController_claimOwnerInvite",
        "summary": "Redeem an invite token to provision a new organization, owner, and default location",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimOrgOwnerInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization workspace provisioned; returns session token"
          },
          "400": {
            "description": "Invalid, expired, or non-owner token"
          }
        },
        "tags": [
          "Invitations & Onboarding"
        ]
      }
    },
    "/organizations/{orgId}/invitations": {
      "post": {
        "operationId": "InvitationsController_createMemberInvite",
        "summary": "Invite a new team member to an organization",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member invitation created successfully"
          },
          "403": {
            "description": "Caller lacks employees:manage permission in organization"
          },
          "409": {
            "description": "Target user is already mapped to this scope"
          }
        },
        "tags": [
          "Invitations & Onboarding"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/auth/claim-member-invite": {
      "post": {
        "operationId": "InvitationsController_claimMemberInvite",
        "summary": "Redeem an invite token to create or link an employee account",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimMemberInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee onboarded; returns session token"
          },
          "400": {
            "description": "Invalid, expired, or non-member token"
          }
        },
        "tags": [
          "Invitations & Onboarding"
        ]
      }
    },
    "/platform/invitations": {
      "post": {
        "operationId": "PlatformController_inviteOrgOwner",
        "summary": "Issue an onboarding invitation for a new Organization Owner",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgOwnerInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization Owner invitation created successfully"
          },
          "403": {
            "description": "Forbidden: Master Admin access required"
          },
          "409": {
            "description": "User with this email already exists"
          }
        },
        "tags": [
          "Platform (Master Admin)"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/platform/organizations": {
      "get": {
        "operationId": "PlatformController_listOrganizations",
        "summary": "List all platform-wide organizations with counts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all platform organizations"
          },
          "403": {
            "description": "Forbidden: Master Admin access required"
          }
        },
        "tags": [
          "Platform (Master Admin)"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Shift Scheduling API",
    "description": "Workforce management & shift scheduling backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.com",
            "description": "Registered user or admin email"
          },
          "password": {
            "type": "string",
            "example": "Admin123!",
            "description": "Account password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "ClaimOrgOwnerInviteDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "a9b8c7...",
            "description": "Cryptographic invite token"
          },
          "fullName": {
            "type": "string",
            "example": "Jordan Belfort"
          },
          "password": {
            "type": "string",
            "example": "SecurePassword123!",
            "minLength": 8
          },
          "organizationName": {
            "type": "string",
            "example": "Apex Logistics Inc."
          },
          "primaryLocationName": {
            "type": "string",
            "example": "Central Warehouse Hub"
          },
          "timezone": {
            "type": "string",
            "example": "America/Toronto",
            "default": "UTC"
          }
        },
        "required": [
          "token",
          "fullName",
          "password",
          "organizationName",
          "primaryLocationName"
        ]
      },
      "CreateMemberInviteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "worker@business.com"
          },
          "roleName": {
            "type": "string",
            "example": "Employee",
            "description": "Existing organization role name"
          },
          "scopeType": {
            "type": "string",
            "enum": [
              "system",
              "organization",
              "location"
            ],
            "example": "location"
          },
          "scopeId": {
            "type": "string",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "positionIds": {
            "example": [
              "3fa85f64-5717-4562-b3fc-2c963f66afa6"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "email",
          "roleName",
          "scopeType",
          "scopeId"
        ]
      },
      "ClaimMemberInviteDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "a9b8c7..."
          },
          "fullName": {
            "type": "string",
            "example": "Alex Mercer"
          },
          "password": {
            "type": "string",
            "example": "SecureWorkerPass123!",
            "minLength": 8
          },
          "phone": {
            "type": "string",
            "example": "+15195550199"
          }
        },
        "required": [
          "token",
          "fullName",
          "password"
        ]
      },
      "CreateOrgOwnerInviteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "owner@newbusiness.com",
            "description": "Prospective business owner email"
          }
        },
        "required": [
          "email"
        ]
      }
    }
  }
}