# src/modules/auth/auth.controller.ts
import {
    Controller,
    Get,
    Post,
    Body,
    Req,
    Res,
    UseGuards,
    HttpCode,
    HttpStatus,
} from '@nestjs/common';
import { AuthGuard } from '@nestjs/passport';
import {
    ApiTags,
    ApiOperation,
    ApiBearerAuth,
    ApiResponse,
} from '@nestjs/swagger';
import { Response } from 'express';
import { AuthService } from './auth.service';
import { LoginDto } from './dto/login.dto';

@ApiTags('Auth')
@Controller('auth')
export class AuthController {
    constructor(private readonly authService: AuthService) { }

    @Post('login')
    @HttpCode(HttpStatus.OK)
    @ApiOperation({
        summary: 'Standard email and password login',
        description: 'Authenticates system user or admin and returns a signed JWT bearer token.',
    })
    @ApiResponse({ status: 200, description: 'Authenticated successfully; returns session token.' })
    @ApiResponse({ status: 401, description: 'Invalid email or password.' })
    async login(@Body() dto: LoginDto) {
        return this.authService.login(dto);
    }

    @Get('google')
    @UseGuards(AuthGuard('google'))
    @ApiOperation({
        summary: 'Trigger Google OAuth login flow',
        description: 'Direct browser link to start Google single sign-on.',
    })
    async googleAuth() {
        // Initiates Google OAuth redirection flow
    }

    @Get('google/callback')
    @UseGuards(AuthGuard('google'))
    @ApiOperation({
        summary: 'Google OAuth callback handler',
        description: 'Receives Google callback code, creates/links user, and redirects to web app.',
  })
  @ApiResponse({ status: 302, description: 'Redirects to frontend auth callback.' })
  async googleAuthRedirect(@Req() req: any, @Res() res: Response) {
      try {
          const { accessToken } = req.user;
      const frontendUrl = process.env.FRONTEND_URL || 'http://localhost:3001';
        return res.redirect(`${frontendUrl}/auth/callback?token=${accessToken}`);
    } catch (error) {
        console.error('OAuth redirect processing failed:', error);
        const frontendUrl = process.env.FRONTEND_URL || 'http://localhost:3001';
        return res.redirect(`${frontendUrl}/login?error=oauth_failed`);
    }
  }

    @Get('me')
    @UseGuards(AuthGuard('jwt'))
    @ApiBearerAuth('JWT-auth')
    @ApiOperation({
        summary: 'Current User Identity',
        description: 'Protected test route to verify your JWT token in Swagger.',
    })
  @ApiResponse({ status: 200, description: 'Current authenticated user profile and roles.' })
  getProfile(@Req() req: any) {
      const userId = req.user.sub || req.user.id || req.user.userId;
      return this.authService.getMe(userId);
  }
}